+++ sky/shell/android/org/domokit/sky/shell/SkyApplication.java
@@ -8,17 +8,21 @@ import android.content.Context;
 import android.util.Log;
 
 import org.chromium.base.BaseChromiumApplication;
-import org.chromium.base.PathUtils;
 import org.chromium.base.library_loader.LibraryLoader;
 import org.chromium.base.library_loader.LibraryProcessType;
 import org.chromium.base.library_loader.ProcessInitException;
+import org.chromium.base.PathUtils;
 import org.chromium.mojo.keyboard.KeyboardServiceImpl;
+import org.chromium.mojo.sensors.SensorServiceImpl;
 import org.chromium.mojo.system.Core;
 import org.chromium.mojo.system.MessagePipeHandle;
 import org.chromium.mojom.activity.Activity;
 import org.chromium.mojom.keyboard.KeyboardService;
+import org.chromium.mojom.media.MediaService;
 import org.chromium.mojom.mojo.NetworkService;
+import org.chromium.mojom.sensors.SensorService;
 import org.domokit.activity.ActivityImpl;
+import org.domokit.media.MediaServiceImpl;
 import org.domokit.oknet.NetworkServiceImpl;
 
 /**
@@ -59,11 +63,10 @@ public class SkyApplication extends BaseChromiumApplication {
       * Override this function to register more services.
       */
     protected void onServiceRegistryAvailable(ServiceRegistry registry) {
-        registry.register(NetworkService.MANAGER.getName(), new ServiceFactory() {
+        registry.register(Activity.MANAGER.getName(), new ServiceFactory() {
             @Override
             public void connectToService(Context context, Core core, MessagePipeHandle pipe) {
-                // TODO(eseidel): Refactor ownership to match other services.
-                new NetworkServiceImpl(context, core, pipe);
+                Activity.MANAGER.bind(new ActivityImpl(), pipe);
             }
         });
 
@@ -74,10 +77,25 @@ public class SkyApplication extends BaseChromiumApplication {
             }
         });
 
-        registry.register(Activity.MANAGER.getName(), new ServiceFactory() {
+        registry.register(MediaService.MANAGER.getName(), new ServiceFactory() {
             @Override
             public void connectToService(Context context, Core core, MessagePipeHandle pipe) {
-                Activity.MANAGER.bind(new ActivityImpl(), pipe);
+                MediaService.MANAGER.bind(new MediaServiceImpl(context, core), pipe);
+            }
+        });
+
+        registry.register(NetworkService.MANAGER.getName(), new ServiceFactory() {
+            @Override
+            public void connectToService(Context context, Core core, MessagePipeHandle pipe) {
+                // TODO(eseidel): Refactor ownership to match other services.
+                new NetworkServiceImpl(context, core, pipe);
+            }
+        });
+
+        registry.register(SensorService.MANAGER.getName(), new ServiceFactory() {
+            @Override
+            public void connectToService(Context context, Core core, MessagePipeHandle pipe) {
+                SensorService.MANAGER.bind(new SensorServiceImpl(context), pipe);
             }
         });
     }
